# Improve performance of logical indexing with BitVectors