Fixes #29418.

The biggest optimization here is that we avoid constantly doing the div/mod operation and indexing into the chunks array by hoisting the intermediate computations into the state tuple.  It's still slower than the straightforward iterative `findall(::BitArray)`, though, so we also specialize `findall` in that case.  Also throws in a cheap trick in case there are lots of falses.

Before:

```julia
julia> using Random, BenchmarkTools; mask = bitrand(10^4); d=rand(10_000);

julia> @btime collect(Base.LogicalIndex($mask));
  65.774 μs (4 allocations: 40.06 KiB)

julia> @btime $d[$mask];
  57.580 μs (4 allocations: 40.25 KiB)
```

After:

```julia
julia> @btime [i for i in Base.LogicalIndex($mask)]; # iteration (unspecialized collect)
  50.238 μs (4 allocations: 39.31 KiB)

julia> @btime collect(Base.LogicalIndex($mask)); # specialized to `findall(mask)`
  42.425 μs (2 allocations: 39.27 KiB)

julia> @btime $d[$mask];
  51.514 μs (4 allocations: 39.31 KiB)
```