@mbauman 2018-10-15
@nanosoldier `runbenchmarks(ALL, vs = ":master")`

Most of this is covered by Nanosoldier — it's just the `collect` method that's missing there.