@nanosoldier 2018-10-15
Something went wrong when running [your job](https://github.com/JuliaLang/julia/pull/29660#issuecomment-430041957):
```
NanosoldierError: failed to run benchmarks against primary commit: failed process: Process(`/home/nanosoldier/workdir/tmpXhooBk/julia -e 'VERSION >= v"0.7.0-DEV.3656" && using Pkg; Pkg.update()'`, ProcessExited(1)) [1]
```
Logs and partial data can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/aff04f959d41bba6e28a9c76aab209e1aaf51e5c/a37505d_vs_57a0347/report.md)
cc @ararslan