@chethega 2018-10-16
Hmm, what do you think of something like the following?
```
julia> Base.@propagate_inbounds function Base.iterate(L::Base.LogicalIndex{Int,<:BitArray}, s=1)
              Bc = L.mask.chunks
              i1,i2 = Base.get_chunks_id(s)
              while true
               @inbounds v = Bc[i1]>>i2
               tz = trailing_zeros(v)
               (tz != 64) && return (s+tz, s+tz+1)
               s = i1<<6 + 1
               i1 += 1
               i2 = 0
               end
              end
```
Idea is to be better on sparser arrays by skipping ahead. Comparison is with the old code, not your branch. Before:
```
julia> d=rand(10_000);
julia> for frac=0.0:0.1:1.0
       m=d.>frac
       @show frac, hash(d[m])
       @btime getindex($d, $m);
       end
(frac, hash(d[m])) = (0.0, 0x469bfa2c4f49fd1a)
  20.245 μs (4 allocations: 78.25 KiB)
(frac, hash(d[m])) = (0.1, 0x77208c630205b24b)
  29.234 μs (4 allocations: 70.19 KiB)
(frac, hash(d[m])) = (0.2, 0x84d7547a33162f3b)
  38.727 μs (4 allocations: 62.31 KiB)
(frac, hash(d[m])) = (0.3, 0x6f1f91da1ccff929)
  48.307 μs (4 allocations: 54.94 KiB)
(frac, hash(d[m])) = (0.4, 0xeba6c7761e5db28c)
  58.146 μs (4 allocations: 46.94 KiB)
(frac, hash(d[m])) = (0.5, 0xf665211ae90f7077)
  61.199 μs (4 allocations: 38.88 KiB)
(frac, hash(d[m])) = (0.6, 0xd01c7407bb028a35)
  55.401 μs (4 allocations: 31.44 KiB)
(frac, hash(d[m])) = (0.7, 0x6796b28c5f486919)
  44.216 μs (4 allocations: 23.69 KiB)
(frac, hash(d[m])) = (0.8, 0x2b35491eb102b91b)
  32.279 μs (3 allocations: 15.80 KiB)
(frac, hash(d[m])) = (0.9, 0xd2982d62ab5cf442)
  21.422 μs (3 allocations: 8.05 KiB)
(frac, hash(d[m])) = (1.0, 0xe26db3b0df0f0fff)
  123.010 ns (3 allocations: 128 bytes)
```
After:
```
julia> for frac=0.0:0.1:1.0
       m=d.>frac
       @show frac, hash(d[m])
       @btime getindex($d, $m);
       end
(frac, hash(d[m])) = (0.0, 0x469bfa2c4f49fd1a)
  53.110 μs (4 allocations: 78.25 KiB)
(frac, hash(d[m])) = (0.1, 0x77208c630205b24b)
  47.897 μs (4 allocations: 70.19 KiB)
(frac, hash(d[m])) = (0.2, 0x84d7547a33162f3b)
  42.767 μs (4 allocations: 62.31 KiB)
(frac, hash(d[m])) = (0.3, 0x6f1f91da1ccff929)
  37.934 μs (4 allocations: 54.94 KiB)
(frac, hash(d[m])) = (0.4, 0xeba6c7761e5db28c)
  32.649 μs (4 allocations: 46.94 KiB)
(frac, hash(d[m])) = (0.5, 0xf665211ae90f7077)
  27.277 μs (4 allocations: 38.88 KiB)
(frac, hash(d[m])) = (0.6, 0xd01c7407bb028a35)
  22.368 μs (4 allocations: 31.44 KiB)
(frac, hash(d[m])) = (0.7, 0x6796b28c5f486919)
  17.374 μs (4 allocations: 23.69 KiB)
(frac, hash(d[m])) = (0.8, 0x2b35491eb102b91b)
  12.071 μs (3 allocations: 15.80 KiB)
(frac, hash(d[m])) = (0.9, 0xd2982d62ab5cf442)
  6.864 μs (3 allocations: 8.05 KiB)
(frac, hash(d[m])) = (1.0, 0xe26db3b0df0f0fff)
  210.157 ns (3 allocations: 128 bytes)
```

Advantage: We don't have as many branch mispredictions and we can skip forward over a bunch of zeros. Disadvantage: we deteriorate hard if most indices are selected. I am almost tempted to insert a density check in the iterator and implement both. But this costs us one well-predicted branch per call.